projects/Alpha-Memory-Restoration/pyproject.toml
[project]
name = "alpha-memory-restoration"
version = "0.1.0"
description = "Add your description here"
requires-python = "==3.12.*"
dependencies = [
    "httpx>=0.28.1",
    "psycopg2-binary>=2.9.11",
    "redis>=7.1.0",
]
